+++ src/entry.h
@@ -216,7 +216,7 @@ class Entry
     DString      exception;   //!< throw specification
     ArgumentList typeConstr;  //!< where clause (C#) for type constraints
     int          bodyLine;    //!< line number of the body in the source
-    int          bodyColumn;  //!< column of the body in the source
+    size_t       bodyColumn;  //!< column of the body in the source
     int          endBodyLine; //!< line number where the definition ends
     int          mGrpId;      //!< member group id
     std::vector<BaseInfo> extends; //!< list of base classes
@@ -224,7 +224,7 @@ class Entry
     std::vector<const SectionInfo*> anchors; //!< list of anchors defined in this entry
     DString	fileName;     //!< file this entry was extracted from
     int		startLine;    //!< start line of entry in the source
-    int		startColumn;  //!< start column of entry in the source
+    size_t      startColumn;  //!< start column of entry in the source
     RefItemVector sli;        //!< special lists (test/todo/bug/deprecated/..) this entry is in
     RequirementRefs rqli;     //!< references to requirements
     SrcLangExt  lang;         //!< programming language in which this entry was found
